+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Methods to consume all gas for a given query or function selector: `givenSelectorConsumeGas(bytes4 selector_)` and `givenQueryConsumeGas(bytes memory query_)`
+
 ### Changed